Update portalGun.ino
I'm not a fan of the Trinket series and the Feather can easily take the place of the Trinker and the LiPoly Backpack so I redesigned my project around that. The MCU is the same so no major changes were needed for the code. The LEDs, Rotary Encoder and Sleep mode all work as expected on my unit.

Reconfigured the power down pin to use Pin #7 as the Adafruit Feather 328 Pdoes not have Pin #8

I also connected a pin to Vol + pin on the FX board to A3 but it seemed to have no effect by triggering it in setup so I ended up having to adjust the WAV files in Audacity. This could also be due to lack of experience with the FX board as this is the first project I've used it it in. I'm leaving the code in because the pin is still connected and everything is now working as expected.

Finally I think it looks better if the front lights only trigger when pressing the button to "fire" the portal gun, so I changed it to do.